+++ src/z3c/relationfield/event.py
@@ -2,6 +2,7 @@
 from z3c.relationfield.interfaces import IHasIncomingRelations
 from z3c.relationfield.interfaces import IHasOutgoingRelations
 from z3c.relationfield.interfaces import IRelation
+from z3c.relationfield.interfaces import IRelationBrokenEvent
 from z3c.relationfield.interfaces import IRelationList
 from z3c.relationfield.interfaces import IRelationValue
 from z3c.relationfield.interfaces import ITemporaryRelationValue
@@ -9,9 +10,17 @@ from zc.relation.interfaces import ICatalog
 from zope import component
 from zope.intid.interfaces import IIntIds
 from zope.event import notify
+from zope.interface import implementer
 from zope.interface import providedBy
 from zope.lifecycleevent import ObjectModifiedEvent
 from zope.schema import getFields
+
+
+@implementer(IRelationBrokenEvent)
+class RelationBrokenEvent(ObjectModifiedEvent):
+    '''
+    Event that is triggered when relation is broken.
+    '''
 
 
 def addRelations(obj, event):
@@ -103,7 +112,7 @@ def breakRelations(event):
     for rel in rels:
         rel.broken(rel.to_path)
         # we also need to update the relations for these objects
-        notify(ObjectModifiedEvent(rel.from_object))
+        notify(RelationBrokenEvent(rel.from_object))
 
 
 def realize_relations(obj):